move tests to single directory
