+++ packages/agents/vickie-bennie/src/lib/tests/bennie.e2e.test.ts
@@ -1,5 +1,5 @@
-import { Bennie } from './Bennie';
-import { container } from './inversify.config';
+import { Bennie } from '../Bennie';
+import { container } from '../inversify.config';
 import {
   Context,
   MachineEvent,
