+++ packages/agents/vickie-bennie/src/lib/tests/gmail.pubsub.e2e.test.ts
@@ -1,6 +1,6 @@
 import { Buffer } from 'buffer';
 import { OfficeService, TYPES } from '@codestrap/developer-foundations-types';
-import { container } from './inversify.config';
+import { container } from '../inversify.config';
 
 if (!process.env.E2E) {
   test.skip('e2e test skipped in default run', () => {
