+++ packages/agents/vickie-bennie/src/lib/tests/processEmailEvent.happyPath.test.ts
@@ -1,23 +1,16 @@
+import { mockProcessEmailEventExecution } from '../__fixtures__/MachineExecutions';
 import {
-  text2ActionTestMachineExecution,
-  machineId,
-} from './__fixtures__/MachineExecutions';
-import { Text2Action } from './Text2Action';
-import { mockEmailResponse } from './__fixtures__/Email';
+  getMockFreeBusyResponse,
+  mockCalendarInsert,
+  mockCalendarList,
+  mockEmailHistoryWithResolution,
+  mockEmailResponse,
+  mockMessageGetResponse,
+  mockMessageGetThreadsResponse,
+} from '../__fixtures__/Email';
+import { Vickie } from '../Vickie';
 
 let counter = 0;
-
-// Mock FoundryClient before any imports that use inversify
-jest.mock('@codestrap/developer-foundations-services-palantir', () => ({
-  ...jest.requireActual('@codestrap/developer-foundations-services-palantir'),
-  createFoundryClient: jest.fn(() => ({
-    // Mock FoundryClient methods as needed
-    someMethod: jest.fn(),
-  })),
-  geminiService: jest.fn(() => {
-    return text2ActionTestMachineExecution.machine;
-  }),
-}));
 
 jest.mock('@codestrap/developer-foundations-utils', () => ({
   ...jest.requireActual('@codestrap/developer-foundations-utils'),
@@ -35,12 +28,12 @@ jest.mock('@codestrap/developer-foundations-data-access-platform', () => ({
         lockOwner?: string,
         lockUntil?: number
       ) => {
-        return text2ActionTestMachineExecution;
+        return mockProcessEmailEventExecution;
       }
     ),
     delete: jest.fn(),
     read: jest.fn((machineExecutionId: string) => {
-      return Promise.resolve(text2ActionTestMachineExecution);
+      return Promise.resolve(mockProcessEmailEventExecution);
     }),
   })),
   makeMemoryRecallDao: jest.fn(() => ({
@@ -125,8 +118,22 @@ jest.mock('googleapis', () => ({
         users: {
           messages: {
             send: jest.fn((request: any) => {
-              console.log(`Gmail mock called with: ${request}`);
+              console.log(`Gmail mock messages.send called with: ${request}`);
               return Promise.resolve(mockEmailResponse);
+            }),
+            list: jest.fn((request: any) => {
+              console.log(`Gmail mock messages.list called with: ${request}`);
+              return Promise.resolve(mockEmailHistoryWithResolution);
+            }),
+            get: jest.fn((request: any) => {
+              console.log(`Gmail mock messages.get called with: ${request}`);
+              return Promise.resolve(mockMessageGetResponse);
+            }),
+          },
+          threads: {
+            get: jest.fn((request: any) => {
+              console.log(`Gmail mock threads.get called with: ${request}`);
+              return Promise.resolve(mockMessageGetThreadsResponse);
             }),
           },
         },
@@ -139,7 +146,28 @@ jest.mock('googleapis', () => ({
         events: {
           insert: jest.fn((request: any) => {
             console.log(`Calendar mock called with: ${request}`);
-            return Promise.resolve({ data: { id: 'mockEventId' } });
+            return Promise.resolve(mockCalendarInsert);
+          }),
+          list: jest.fn((params: any) => {
+            console.log(
+              `Calendar mock events.list called with: ${JSON.stringify(params)}`
+            );
+            return Promise.resolve(mockCalendarList);
+          }),
+        },
+        /* ---------- freebusy.query mock ---------- */
+        freebusy: {
+          query: jest.fn((params: any) => {
+            console.log(
+              `Calendar mock freebusy.query called with: ${JSON.stringify(
+                params
+              )}`
+            );
+            const mockResponse = getMockFreeBusyResponse(
+              params.requestBody.timeMin,
+              params.requestBody.timeMax
+            );
+            return Promise.resolve(mockResponse);
           }),
         },
       };
@@ -183,25 +211,19 @@ jest.mock('googleapis', () => ({
   },
 }));
 
-describe('testing Text2Action', () => {
+describe('testing Vickie', () => {
   afterAll(() => {
     jest.clearAllMocks();
   });
 
-  it('it should rehydrate an existing execution and return pause', async () => {
-    const solution = {
-      input: '', //not relevant for this
-      id: machineId || '',
-      plan: '', //not relevant for retrieving an execution
-    };
-
-    const t2a = new Text2Action();
-    const result = await t2a.upsertState(undefined, true, machineId);
-    const state = JSON.parse(result.state!);
-    // TODO make this test better. Currently we are returning the mock value which
-    // does not reflect the updates which should return the success state
-    expect(state.value).toBe('sendEmail|1');
-    expect(state.context.stack).toHaveLength(1);
-    expect(state.context.stack[0]).toBe('sendEmail|1');
-  }, 30000);
+  it('it handle a mock event using processEmailEvent when a resolution is found', async () => {
+    const vickie = new Vickie();
+    const result = await vickie.processEmailEvent(
+      'eyJlbWFpbEFkZHJlc3MiOiJkc21pbGV5QGNvZGVzdHJhcC5tZSIsImhpc3RvcnlJZCI6MTc5MDUxMn0=',
+      '2025-07-22T20:43:55.184Z'
+    );
+    expect(result).toBeDefined();
+    expect(result.message).toBeDefined();
+    expect(result.status).toBe(200);
+  }, 120000);
 });
