+++ packages/agents/vickie-bennie/src/lib/tests/researchAssistant.e2e.test.ts
